pyproject.toml
[tool.poetry]
name = "gendiff"
version = "0.1.0"
description = ""
authors = ["Ivansergee <klimentievis@inbox.ru>"]
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
flake8 = "^4.0.1"
PyYAML = "^6.0"
asciinema = "^2.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
